Move Button stories to app folder
